Move models to respective namespace.
