+++ src/Models/Position.cs
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Gauge-Dotnet.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace Gauge.Dotnet
+namespace Gauge.Dotnet.Models
 {
     public class Position
     {
